Fix Bug on  #4370 - Upgradae to WP 3.6.1 caused errors
The issue is explained here
http://php.net/manual/en/language.references.pass.php

Note: There is no reference sign on a function call - only on
function definitions. Function definitions alone are enough to
correctly pass the argument by reference. As of PHP 5.3.0, you
will get a warning saying that "call-time pass-by-reference" is
deprecated when you use & in foo(&$a);. And as of PHP 5.4.0,
call-time pass-by-reference was removed, so using it will raise a
fatal error.