Remove html tags from the_excerpt
the_excerpt was allowing html to be output
this caused the page to render with unclosed tags
in situations where the excerpt reached it's
character limit before closing the tag. Unclosed
tags cause the page to render improperly usual
frontend que was sidebar not appearing in the
side of the page, rather wrapping under content.